StevenzPlayground posting
My work in progress. Full of printfs and such. Gates are working, except for Enable and Sustain modes.Contains Guy's work on averagingand filtering for slider values. Has enabled time sliders on recall and latest catch-up scheme for sliders. A mess, but it compiles and runs well.Steup for printf to IDM.