Update main.c
FIxed bug in Calibration for Stage Address 2. Seems like the value for that control was being inconsistently measured (it was the last of 8 measurements read) . Problem was non-deterministic end to the Calibrate function. Adding a return to finish function  after writing values instead of waiting for release of Advance key solved issue